# server/api/api/swagger/swagger.yaml
swagger: '2.0'
info:
  description: 이 문서는 스마트온실 UI를 위한 API 문서입니다.
  version: "0.3.0"
  title: Smart Greenhouse API
  contact:
    email: dev.jinong@gmail.com
host: "localhost:8081"
basePath: /common/v1
...
schemes:
  - http
consumes:
  - application/octet-stream
  - application/json
produces:
  - application/json  
securityDefinitions:
  Bearer:
    type: apiKey
    name: authorization
    in: header
paths:
...
  /alarm/config/{identifier}:
    x-swagger-router-controller: jnalarm
    get:
      tags: 
      - config
      description: 설정정보를 조회한다
      operationId: getConfigInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
      responses:
        "200":
          description: Success. 
          schema:
            $ref: '#/definitions/AlarmConfiguration'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"
    put:
      tags:
      - config
      description: 설정정보를 수정한다
      operationId: updateConfigInfo
      produces:
        - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
        - in: body
          name: body
          description: 설정정보
          required: true
          schema:
              $ref: '#/definitions/AlarmConfiguration'  
      responses:
        "200":
          description: success
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"          
  /alarm/threshold/{identifier}:
    x-swagger-router-controller: jnalarm
    get:
      tags: 
      - threshhold
      description: 임계값 정보를 조회한다
      operationId: getThresholdInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/AlarmThreshold'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"
    put:
      tags:
      - threshhold
      description: 임계값 정보를 수정한다
      operationId: updateThresholdInfo
      produces:
        - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
        - in: body
          name: body
          description: 임계값 수정정보
          required: true
          schema:
            $ref: '#/definitions/AlarmThreshold'
      responses:
        "200":
          description: success 
          schema:
            type: string
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"
  /alarm/history/{identifier}:
    x-swagger-router-controller: jnalarm
    get:
      tags: 
      - history
      description: 이상값 로그정보를 조회한다
      operationId: getHistoryInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string      
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/AlarmEvent'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"
  /alarm/observation/{identifier}:
    x-swagger-router-controller: jnalarm
    get:
      tags: 
      - observation
      description: 현재 센서값 정보를 조회한다
      operationId: getObservationInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/AlarmObservation'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"            
  /alarm/app/{identifier}:          
    x-swagger-router-controller: jnalarm
    post:
      tags:
        - app
      description: 앱정보(앱버전, fcm key) 를 입력한다
      operationId: updateAppInfo
      produces:
        - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
        - in: body
          name: body
          description: 앱정보 
          required: true
          schema:
              $ref: '#/definitions/AlarmAppInfo'
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "405" : 
          description: Bad RequestMethod Not Allowed
          schema:
            type: string
        "500" :
          description: Internal Server Error
          schema:
            type: string
  /alarm/version/{identifier}:            
    x-swagger-router-controller: jnalarm         
    get:
      tags: 
      - version
      description: app 버전정보를 조회한다
      operationId: getVersionInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string  
      responses:
        "200":
          description: Success
          schema:
            $ref: '#/definitions/AlarmVersion'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"                  
  /alarm/status/{identifier}:            
    x-swagger-router-controller: jnalarm         
    get:
      tags: 
      - status
      description: 알람정보 상태를 조회한다
      operationId: getAlarmStatusInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string  
      responses:
        "200":
          description: Success
          schema:
            type: array
            items:
              $ref: '#/definitions/AlarmEvent'
        default:
          description: Error
          schema:
            $ref: "#/definitions/error"  
    put:
      tags: 
      - status
      description: 알람 지연시간을 설정한다. 다른것도 설정할 수 있는 형태이지만 내부적으로 지연시간만을 반영한다. delay를 빈 문자열로 하면 알람을 전송하지 않는 것으로 설정한다. 지연시간은 일회성으로 한번 지연시간이 초과하면 이후부터는 디폴트 값으로 지연된다.
      operationId: setAlarmStatusInfo
      produces:
      - application/json
      parameters:
        - name: identifier
          in: path
          description: 핸드폰 번호로 중간에 하이픈이 없는 형태로 한다. (ex. 01012345678)
          required: true
          type: string
        - in: body
          name: body
          description: 앱정보 
          required: true
          schema:
              $ref: '#/definitions/AlarmEvent'
      responses:
        "200":
          description: Success
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "405" : 
          description: Bad RequestMethod Not Allowed
          schema:
            type: string
        "500" :
          description: Internal Server Error
          schema:
            type: string
...
  /feed/download:
    x-swagger-router-controller: knu
    get:
      tags:
        - knu_feed
      description: 경북대 사료정보(수분, ph) 다운로드
      operationId: getKnuDownload
      produces:
        - text/csv
      responses:
        '200':
          description: successful operation
          schema:
            type: file
        '405':
          description: Invalid input
definitions:
...
  AlarmThreshold:
    type: object
    properties:
      'pname':
        type: string
      'used':
        type: boolean
      'thresholds':
        type: array
        items:
          $ref: '#/definitions/AlarmOneThreshold'
  AlarmOneThreshold:
    type: object
    properties:
      name:
        type: string
      value:
        type: number
  AlarmObservation:
    type: object
    properties:
      name:
        type: string
      unit:
        type: string
      value:
        type: number
      obstime:
        type: string
      isgood:
        type: boolean
  AlarmEvent:
    type: object
    properties:
      id:
        type: number
      name:
        type: string
      occured:
        type: string
      updated:
        type: string
      severity:
        type: number
  AlarmConfiguration:
    type: object
    properties:
      defdelay:
        type: integer
        example: 10
        description: 분단위. 디폴트 지연시간.
      method:
        type: string
        example: 'V'
        description: V:진동 N:사용안함 R:소리 M:문자메세지 K:소리/진동
  AlarmAppInfo:
    type: object
    properties:
      appversion:
        type: string
        example: '0.1'
      fcmkey:
        type: string
        example: 'fuevENfVbwY:APA91bFBz_AB2lkr940n6kIYnkzEqqjs4eyJKtCy0AFK67m0X3K9v46ekUagEb'
  AlarmVersion:        
    type: object
    properties:
      newest_app_version:
        type: string
        example: '0.3'
      minimum_support_app_version:
        type: string
        example: '0.2'
...
  error:
    required:
      - message
    properties:
      message:
        type: string
        example: 'false'